+++ backend/functions/src/api/get-thesaurus-data.js
@@ -1,32 +1,4 @@
-const https = require('https')
-const functions = require('firebase-functions')
-const admin = require('firebase-admin')
-
-
-// Set this with: firebase functions:config:set thesaurusapi.key="yourkey"
-// Register for an API key at https://dictionaryapi.com/
-const thesaurusAPIKey = functions.config().thesaurusapi.key
-
-function getThesaurusResponse(word, cb) {
-  const url = `https://dictionaryapi.com/api/v3/references/thesaurus/json/${word}?key=${thesaurusAPIKey}`
-
-  https.get(url, res => {
-    res.setEncoding("utf8")
-    let body = ""
-    res.on("data", data => {
-      body += data
-    })
-    res.on("end", () => {
-      try {
-        body = JSON.parse(body)
-      } catch {
-        cb('Unable to decode JSON: ' + body)
-        return
-      }
-      cb(null, body)
-    });
-  });
-}
+const getThesaurusResponse = require('../util/thesaurus')
 
 module.exports = function sendThesaurusData(req, res) {
   /*