+++ simply-synonyms/content/content.js
@@ -6,7 +6,12 @@ let headers = {} // Request headers (for auth)
 
 function getSynonyms(word) {
   word = word.trim()
-  if (loggedIn) fetch(`https://us-central1-simply-synonyms-api.cloudfunctions.net/api/update-user-stats`, { headers }) // Increment user's synonym counters
+  if (loggedIn) fetch(`https://us-central1-simply-synonyms-api.cloudfunctions.net/api/update-user-stats`, { headers })// Increment user's synonym counters
+    .then(({ status }) => {
+      if (status === 401) {
+        chrome.runtime.sendMessage(null, { action: 'refreshIdToken' }, {}, setHeaders) // Check for a token refresh and reset headers when fetch is unauthorized
+      }
+    })
   return fetch(`https://us-central1-simply-synonyms-api.cloudfunctions.net/api/get-thesaurus-data?word=${word}`)
     .then(response => response.json())
     .then(data => {
@@ -200,6 +205,12 @@ function addExtension() {
   })
 }
 
+function setHeaders(t) {
+  if (!t) return
+  headers = new Headers({
+    'Authorization': `Bearer ${t}`
+  })
+}
 
 chrome.storage.local.get(['option_popupDisabled', 'option_onlyEditableText'], (result) => {
   options = result
@@ -208,9 +219,7 @@ chrome.storage.local.get(['option_popupDisabled', 'option_onlyEditableText'], (r
 
 chrome.storage.local.get(['idToken'], ({ idToken }) => {
   loggedIn = !!idToken
-  if (idToken) headers = new Headers({
-    'Authorization': `Bearer ${idToken}`
-  })
+  setHeaders(idToken)
 })
 
 // chrome.runtime.onMessage.addListener((message, sender, respond) => {