+++ changelog.md
@@ -4,7 +4,7 @@
  + Migrated to Firebase
  + Sign in with Google Chrome identity
  + Welcome messages in extension popup
- + Counts how many synonyms users look up per week
+ + Counts how many synonyms users look up per week, year and account
  + Mailgun email system
  + Welcome email on first sign in
  + Optional weekly stats and synonyms email